AI-Agents/Agent-05-liteLLM/pyproject.toml
[project]
name = "agent-05-litellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.72.6.post1",
    "python-dotenv>=1.1.0",
]
